rename BindFooter
