+++ src/components/Template/BaseLayout/index.js
@@ -1,10 +1,10 @@
 import BindNavigation from '@/components/Organism/BindNavigation/index.vue'
-import BindFooter from '@/components/Organism/BindFooter/index.vue'
+import TheBindFooter from '~/components/Organism/TheBindFooter/index.vue'
 
 export default {
   name: 'BaseLayout',
   components: {
     BindNavigation,
-    BindFooter,
+    TheBindFooter,
   },
 }
